pyproject.toml
[tool.poetry]
name = "oucwebproject"
version = "0.1.0"
description = ""
authors = ["Lingxi <2863878907@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
selenium = "^4.24.0"
pillow = "^10.4.0"
pytesseract = "^0.3.13"
opencv-python = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
